cleanup e2e tests - part I
